Fix version in config file generated by goal cmd (#447)
Before the fix 'goal network create' generated config.json
without 'Version' field. That forces config loader to assume Version=0
and start migrations.
If BroadcastConnectionsLimit set to 0 (special value) it is overwritten by -1
that is a new parameter introduced in config v4.
BroadcastConnectionsLimit does not exist in v3 with assumed value 0 in migration.